Added Bill Summary
Added method to return a list of bill summaries based on account number.
Added integration test for this.
Fixed bugs revealed by unit test.
