+++ src/main/java/org/drip/services/AccountService.java
@@ -1,6 +1,9 @@
 package org.drip.services;
 
+import java.util.List;
+
 import org.drip.model.Account;
+import org.drip.model.BillSummary;
 import org.drip.model.Payment;
 
 public interface AccountService {
@@ -8,5 +11,7 @@ public interface AccountService {
 	Account getAccount(int customerId);
 	
 	Payment getPayment(String accountId);
+
+	List<BillSummary> getBillSummaries(String accountNumber);
 	
 }
